backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
yt-dlp==2025.7.21
psycopg2-binary==2.9.9
pandas==2.3.0
openpyxl==3.1.5
python-multipart==0.0.9
python-dotenv==1.1.1
openai==1.96.1
sqlalchemy==2.0.41
pydantic==2.10.1
alembic==1.16.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
google-generativeai==0.8.3
markdown2
reportlab
requests==2.31.0
beautifulsoup4==4.12.3
